requirements.txt
boto3
opencv-python
#pillow
numpy
#pytesseract
easyocr
jsonschema
sagemaker
python-dotenv
